working 4/23

+++ androidFW/src/org/kotemaru/android/fw/util/WindowUtil.java
@@ -65,7 +65,7 @@ public class WindowUtil {
 	}
 
 
-	public static int sp2px(Context context, int sp) {
+	public static int sp2px(Context context, float sp) {
 		final float scale = context.getResources().getDisplayMetrics().scaledDensity;
 		return (int) (sp * scale + 0.5f);
 	}
